requirement.txt.txt
streamlit
opencv-python
numpy
scikit-learn
Pillow
matplotlib
seaborn
